requirements.txt
requests
pyyaml
beautifulsoup4